chore: drop old canvas renderer
